adopted new format for all relevant data loaders other than hcl, mca and tms
