+++ sfaira/data/dataloaders/loaders/d10_1038_s41591_019_0468_5/human_lung_2019_dropseq_braga_001.py
@@ -54,11 +54,11 @@ class Dataset(DatasetBase):
         }
 
     def _load(self, fn=None):
-        if fn is None:
-            fn = [
-                os.path.join(self.path, "human", "lung", "GSE130148_raw_counts.csv.gz"),
-                os.path.join(self.path, "human", "lung", "GSE130148_barcodes_cell_types.txt.gz"),
-            ]
+        base_path = os.path.join(self.path, "human", "placenta")
+        fn = [
+            os.path.join(base_path, "GSE130148_raw_counts.csv.gz"),
+            os.path.join(base_path, "GSE130148_barcodes_cell_types.txt.gz"),
+        ]
         self.adata = anndata.read_csv(fn[0]).T
         self.adata.obs = pd.read_csv(fn[1], sep="\t", index_col=0)
