+++ sfaira/data/dataloaders/loaders/d10_1101_2020_03_13_991455/human_lung_2020_10x_lukassen_001.py
@@ -4,23 +4,29 @@ from typing import Union
 import numpy as np
 import scipy.sparse
 
-from sfaira.data import DatasetBase
+from sfaira.data import DatasetBaseGroupLoadingManyFiles
+
+SAMPLE_FNS = [
+    "lukassen20_lung_orig.processed.h5ad",
+    "lukassen20_airway_orig.processed.h5ad"
+]
 
 
-class Dataset(DatasetBase):
+class Dataset(DatasetBaseGroupLoadingManyFiles):
 
     def __init__(
             self,
+            sample_fn: str,
             path: Union[str, None] = None,
             meta_path: Union[str, None] = None,
             cache_path: Union[str, None] = None,
             **kwargs
     ):
-        super().__init__(path=path, meta_path=meta_path, cache_path=cache_path, **kwargs)
-        self.id = "human_lung_2020_10x_lukassen_001_10.1101/2020.03.13.991455"
+        super().__init__(sample_fn=sample_fn, path=path, meta_path=meta_path, cache_path=cache_path, **kwargs)
+        self.id = f"human_lung_2020_10x_lukassen_{str(SAMPLE_FNS.index(self.sample_fn)).zfill(3)}_" \
+                  f"10.1101/2020.03.13.991455"
 
-        self.download = "https://covid19.cog.sanger.ac.uk/lukassen20_lung_orig.processed.h5ad"
-        self.download_meta = None
+        self.download = f"https://covid19.cog.sanger.ac.uk/{self.sample_fn}"
 
         self.author = "Eils"
         self.doi = "10.1101/2020.03.13.991455"
@@ -36,23 +42,43 @@ class Dataset(DatasetBase):
 
         self.obs_key_cellontology_original = "CellType"
 
-        self.class_maps = {
-            "0": {
-                "Ciliated": "Multiciliated lineage",
-                "Endothelial": "1_Endothelial",
-                "AT2": "AT2",
-                "LymphaticEndothelium": "Lymphatic EC",
-                "Fibroblasts": "2_Fibroblast lineage",
-                "Club": "Secretory",
-                "Immuno_TCells": "T cell lineage",
-                "Immuno_Monocytes": "Monocytes",
-                "AT1": "AT1"
-            },
-        }
+        if self.sample_fn == "lukassen20_lung_orig.processed.h5ad":
+            self.class_maps = {
+                "0": {
+                    "AT1": "AT1",
+                    "AT2": "AT2",
+                    "Ciliated": "Multiciliated lineage",
+                    "Club": "Secretory",
+                    "Endothelial": "1_Endothelial",
+                    "Fibroblasts": "2_Fibroblast lineage",
+                    "Immuno_TCells": "T cell lineage",
+                    "Immuno_Monocytes": "Monocytes",
+                    "LymphaticEndothelium": "Lymphatic EC",
+                }
+            }
+        else:
+            self.class_maps = {
+                "0": {
+                    "Basal_Mitotic": "Basal",
+                    "Basal1": "Basal",
+                    "Basal2": "Basal",
+                    "Basal3": "Basal",
+                    "Ciliated1": "Multiciliated lineage",
+                    "Ciliated2": "Multiciliated lineage",
+                    "Club": "Secretory",
+                    "Fibroblast": "2_Fibroblast lineage",
+                    "FOXN4": "Rare",
+                    "Ionocyte": "Rare",
+                    "Goblet": "Secretory",
+                    "Secretory3": "Secretory",
+                    "Secretory2": "Secretory",
+                    "Secretory1": "Secretory",
+                },
+            }
 
     def _load(self, fn=None):
-        if fn is None:
-            fn = os.path.join(self.path, "human", "lung", "lukassen20_lung_orig.processed.h5ad")
+        base_path = os.path.join(self.path, "human", "lung")
+        fn = os.path.join(base_path, self.sample_fn)
         self.adata = anndata.read(fn)
         self.adata.X = np.expm1(self.adata.X)
         self.adata.X = self.adata.X.multiply(scipy.sparse.csc_matrix(self.adata.obs["nCount_RNA"].values[:, None]))\
