+++ sfaira/data/templates/dataloaders/many_samples_many_files_streamlined/your_dataset_file.py
@@ -21,8 +21,8 @@ class Dataset(DatasetBaseGroupLoadingManyFiles):
         super().__init__(
             sample_fn=sample_fn,
             path=path, meta_path=meta_path, cache_path=cache_path, **kwargs)
-
+        self.id = f"sth_{str(SAMPLE_FNS.index(sample_fn)).zfill(3)}_doi"  # ToDo: Index the Dataset ID by the file.
         # ToDo Add you meta data here.
 
-    def _load_any_object(self, fn):
-        pass  # ToDo: load file fn into self.adata.
+    def _load_any_object(self, fn=None):
+        pass  # ToDo: load file fn into self.adata, using self.sample_fn, ie the current sample file.
